refactor(tests): split shader param translation tests and surface animator harness
